修改

+++ app/src/main/java/com/cmoney/kolfanci/service/media/RecorderAndPlayer.kt
@@ -46,25 +46,25 @@ interface RecorderAndPlayer {
      * 獲得播放的總時長
      * @return 播放時長（毫秒）
      */
-    fun getPlayingDuration(): Int
+    fun getPlayingDuration(): Long
 
     /**
      * 獲得當前播放的秒數
      * @return 播放當前毫秒數的狀態流
      */
-    fun getPlayingCurrentMilliseconds(): StateFlow<Int>
+    fun getPlayingCurrentMilliseconds(): StateFlow<Long>
 
     /**
      * 獲得當前錄音的秒數
      * @return 錄音當前毫秒數的狀態流
      */
-    fun getRecordingCurrentMilliseconds(): StateFlow<Int>
+    fun getRecordingCurrentMilliseconds(): StateFlow<Long>
 
     /**
      * 獲得錄音的總時長
      * @return 錄音時長（毫秒）
      */
-    fun getRecordingDuration(): Int
+    fun getRecordingDuration(): Long
     fun getFileUri(): Uri?
 
     /**
